// generated/src/app/api-models/cap-poolcalc-work.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class CapPoolcalcWork {
  seqCcalcId: number;
  seqCapPoolcalcWork: number;
  seqCapEntityWork: number;
  seqCapEligWork: number;
  seqCapPoolId: number;
  capMonth: Date;
  rateDet1: string;
  rateDet2: string;
  rateDet3: string;
  pmpmAlloc: number;
  pmpmWith: number;
  seqAgeBandDtl: number;
  effectiveDate: Date;
  termDate: Date;
  pmpmPostRule: string;
  pmpmCoCode: string;
  pmpmGlRef: string;
  currentRetroFlag: string;
  addChgTermFlag: string;
  seqMembId: number;
  seqGroupId: number;
  planCode: string;
  age: number;
  gender: string;
  rptMonthCap: Date;
  seqCapRateHdr: number;
  capRate: number;
  positiveAmtFlag: string;
  retroRateFlag: string;
  seqIncentiveRule: number;
  seqProvContract: number;
  accessIncentiveAmount: number;
  accessIncentivePercent: number;
  accessCapApply: string;
  incentiveRetro: string;
}